2x speedup and center grid (and clean trailing whitespace)
the kinemetics take advantage of symmetry and calculate everything as if
it happend on one side of the workspace. This commit takes advantage of
that by only calculating points on half the workspace and then using
those points to draw lines across the entire workspace.

In doing this, I also center the grid, so that there is always a point
at 0,0

I also cleaned some trailing whitespace that I had acciently introduced,
and in the process cleaned trailing whitespace that was already there.