fix NoneType error
I found that there was incorrect indentation in kinematics.py which made
the chain length calculations and return statement be something inside
the loop instead of outside the loop.

This meant that if the errors were small enough to begin with, instead
of returning chain lengths based on the first guess values, nothing
would be returned.

Similarly, if it took more than 10 tries to find the chain length,
nothing would be returned as well (and no error output). Now there is
error output.