reduce the insance accuracy requirements
besides eliminating some failures that will cause the machine to head to
the center of the workspace, this will also speed up calculations

the forward routine was requiring accuracy of 0.000001mm in the chain
length.

the hardware has a theoretical accuracy of 0.0078mm in the chain length.

This reduces the required accuracy to 0.01mm for the chain length
calculations in forward

In inverse, it was requiring an accuracy of 0.001mm, the design target
is 0.39mm (1/64"), so reduce the accuracy requirement to 0.01mm

This will make things faster, but will also eliminate faiures that
happen at some points. I did a check with a grid size of 1mm and the
following points failed due to this accuracy requirement

508.000011373 -525.000467924
1101.00223304 476.006380171
276.000984612 -400.002655064
569.999343355 -556.003109189
473.000022234 -507.000103024