Fix issue with random circles being cut
There was an issue with random circles being cut as full circles instead of partial circles, the solution was to turn off the automatic truncating of the number of digits being sent to the firmware.

Unfortuantely becuase the settings file is only created the FIRST time ground control is installed the only to make the change for existing Ground Control installs is to force the setting value to false every time Ground Control is launched, so for the next few weeks to use the very usefull feature of truncating the number of digits sent to the machine you will have to go into settings and manually enable it each time GC launches. Once most peoples installs have been fixed, we'll turn off the auto disable on launch part.