Modify Simulation for Chain Slippage
Simulate what it would look like if the chain position is off by a
specified number of links.

This error is not symmetrical so it required running the simulation on the
full work area, not just the left side and mirroring for the right.  This
does result in taking twice as long to generate a simulation, but I think
it is worth for the added capability.