Ignore decimal places in gotoLine popup
In gotoLine popup, if the user enters a number with decimal places such as "123.45", the entire number is ignored. This is because the try fails when converting this string directly to an integer. By changing it to first convert to float, then to int, the integer portion is preserved and the decimal places are ignored.