Enable keyboard input for NotificationPopup and TouchNumberInput popups
When accepting notifications or entering numbers, it would be nice to use a real keyboard, especially on computers without touch screens.  The GcodeCanvas class is already using keyboard input via kivy request_keyboard for zooming in and out, so adding request_keyboard to the popups breaks the callback in GcodeCanvas.  However, kivy supports multiple callbacks for the on_key_down event, with the most recently bound callback being called first, and when a callback returns true, it prevents the event from being sent to the rest of the callbacks.  Thus, by binding and unbinding to the on_key_down event, we can allow keyboard input for modal dialogs and revert events back to the original GcodeCanvas when the dialogs are closed.  The added "global_variables" module is not the most elegant solution, but it gets the job done for now.  Tested with position numeric input dialog open, with notification dialog open on top of that, and keyboard input worked correctly for both dialogs and GcodeCanvas.