Fix Popup Numeric Display (clean)
The number displayed after closing a popup (touchNumberInput or z-Axis Popup) is the text version of what was entered.  If the user enters "000" the number displayed will be "000".  If the user enters ".1" the number displayed will be ".1".
By converting the number to a float and back to a string, we can display the number in standard format.  If the user enters "000" the number displayed will be "0.0".  If the user enters ".1" the number displayed will be "0.1".  (Now trying as clean update to original master.)