Allow windowed mode
Under the settings for kivy, there is an option for Fullscreen, which allows the user to choose windowed or fullscreen mode.  But Ground Control forces fullscreen mode regardless of this setting by calling window.maximize() in the build() function of main.py.  This change removes window.maximize() to allow the user to choose windowed or fullscreen mode.

A side effect of this change is that the Gcode canvas is not centered in the window.  This is because the call to window.maximize() triggered the on_resize() callback.  Now that the window.maximize() call is removed, the callback does not occur at initialization.  To fix this, a call to self.centerCanvas() is added to the initialize() function of gcodeCanvas.py.