Add a switch to step 7 to enable z-axis
In step 7 of the calibration process it was unclear that the z-axis needed to be inabled in settings, and dificult to do so.

This PR updates the text to make it  more clear that it needs to be enabled, and adds a swtich to let you enable it right from step 7.

A fix for issue #407