preserve X and Y positions
If a gcode line does not contain either an X or a Y, we need the previous value to correctly render the cursor during the single-step preview. Previously if X or Y was absent, we used self.gcodecanvas.positionIndicator.pos[] values, but these are over-written by the real-timeposition of the router.