Change the default encoder steps from 8148 to 8114
As identified in this forum thread the correct (or more correct) value for the encoderSteps variable is 8114.

Because this update effects a value stored in the users settings file not only does the default value in ground control need to be changed, but also the users .ini file needs to be updated.

The PR will detect if a user is using the old value and if they are the value will be updated and a popup will let them know that the change has been made.