Ensure Machine is Ready; Log Inc Messages Immediately
There was no check for the 'ok' command from the firmware
before sending the first message.  This is likely why PR #405
and commit 6c934aad9b39438111242fae73e5617cc4321377 which sent
a blank line on startup helped, becuase only the first command
could be sent without an ok.  Removed the kludge of a blank
command and it seems to work well.

This may also fix the issues reported in this post:
https://forums.maslowcnc.com/t/unable-to-execute-command-machine-settings-not-yet-received/1517

Also added a new child class of Queue which logs the messages
to the logger on put rather than waiting for the process to get
them out of the queue. The get method was resulting in non-
sequential incoming and outgoing messages in the log because
outgoing messages were logged immediately.  This makes the log
much more helpful.