Add a sanity check to make sure that the values entered are realistic in triangular calibration
We have had issues with this text box accepting any number. If the number is not a realistic number say (1,000,000) mm the machine will still do it's best to find a calibration which works for that number resulting in nonsense calibration (like a negitive rotation radius). This pull request adds a sanity check which limits the entered value to +-300mm of the expected 1905mm distance