fix vertical placement slider logic
With the vertical placement slider of the Triangular Calibration step at the top of its range (in the neighborhood of +388mm judging by the log), after making the cuts, the sled tries to move 500mm above that level. That would be above the top of a 1220mm sheet.
If the cut level is above Y=0, make the sled move in the negative Y directionn after the cut instead?
The move is currently 500mm, that gets pretty close to the edge when starting from the middle. Lowering that to 400...