Disable command to zero the axes in Step1 'Begin' button
I think the issue I was seeing was caused by the 'Begin' button in Step1 - it executes gcode "B06 L0 R0" which tells both axes that they are at the first link of the chain. There is a comment "This prevents strange behavior when rotating each sprocket to 12:00".
With the present code, clicking this button clobbers your chain calibration whether that was what you intended or not.
 I believe recent changes to kinematics and the chain calibration procedure make this 'Begin' button behaviour unnecessary.