Simplify How Settings Are Handled
Make a truly single source for all attributes about all settings.
Added some helper functions as well.

Change on_change_config to callback.  Prior function was only called
if setting was changed on the config page.  This way it is actually
called when a setting is changed no matter where.