A proposed fix for #306
This is a proposed fix for the issue where the machine will not move again after pressing the stop button.

The solution is to indicate that when a stop has been sent it will free up the entire buffer. As soon as the OK message is recieved to indicate that the stop command has been processed the entire buffer is marked as free.

One more change I would like to make if we like this solution is to change all of the hard coded 256s in there to be a #define so it is more clear why that number is used and the code can be easily modified for different buffer sizes.

Closes #306