Add a try...except catch to velocity calculation
Sometimes this line is failing with a division by zero error...probably because time() is the same as self.lasttime

This could be because the value of last time didn't update properly somewhere or because the function was called back to back and something went wrong on the OS side.

The proposed solution here is to just ignore it when that happens and don't update the velocity for that command.

It's not super elegant.