rename 'zAxisAuto' to 'spindleAutomate'
PR #458 added a setting to enable/disable spindle control. The setting was named 'zAxisAuto' which seems confusing as it is being used as an on/off control for spindle power. Change the name to 'spindleAutomation' to better reflect its use. Note that the change in GC will match an accompanying PR in Firmware.