Correct rendering error wnen the arc to render is a full circle
This addresses issue #672:
If the starting XY and ending XY of a G02 or G03 code are identical, the circle will not be rendered on the gcodeCanvas. When that same line is sent to the firmware, it will be correctly processed.
Fixed the previous trig error, arcLen is in radians instead of degrees...