Make motor stall warning not trigger on pull chain tight
This disables the new feature which will detect motor stalls during the pull chain tight step when we want to stall the motor in a controled way.

On an side note, it was so cool to see the motor stall warning pop up! I'm really excited about this new feature, I think it's going save a lot of chips :+1: