Increase max feedrate to 800mm/min
We lowered the maximum feedrate from 1000 to 700 a while back in preperation for this batch of machines going out. It was a really conservative number to make sure everything went smoothly, but now that we are through the thick of getting them up and running I would like to start inching that back up. This PR ups it to 800 which I've been running for a while now with no issues